Restructure the whole project
