+++ public/admin/item/logic/item_delete.php
@@ -1,5 +1,5 @@
 <?php
-include_once __DIR__ . '/../../includes/connect-db.php';
+include_once __DIR__ . '/../../../includes/connect-db.php';
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     if (isset($_POST['iditem'])) {
@@ -9,14 +9,14 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             $stmt = $pdo->prepare("DELETE FROM item WHERE iditem = :iditem");
             $stmt->bindParam(':iditem', $iditem);
             $stmt->execute();
-            header("Location: ../../item.php");
+            header("Location: ../../item/index.php");
             exit();
         } catch (PDOException $e) {
-            header("Location: ../../item.php?error=1");
+            header("Location: ../../item/index.php?error=1");
             exit();
         }
     } else {
-        header("Location: ../../item.php?error=1");
+        header("Location: ../../item/index.php?error=1");
         exit();
     }
 }
