+++ public/admin/users/logic/user_add.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-include_once '../../includes/connect-db.php';
+include_once '../../../includes/connect-db.php';
 
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $l_name = trim($_POST['l_name']);
@@ -32,7 +32,7 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $_SESSION['message_type'] = "error";
     }
 
-    header("Location: ../../user.php");
+    header("Location: ../../users");
     exit();
 }
 ?>
