+++ public/admin/users/logic/user_admin_edit.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-include_once '../../includes/connect-db.php';
+include_once '../../../includes/connect-db.php';
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['update_user'])) {
     $user_id = $_POST['user_id'];
@@ -29,6 +29,6 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['update_user'])) {
 } else {
     $_SESSION['message'] = "Invalid request!";
     $_SESSION['message_type'] = "error";
-    header("Location: ../../user.php");
+    header("Location: ../../users");
     exit();
 }
