--- public/js/user_dashboard.js
@@ -1,11 +1,0 @@
-$(document).ready(function () {
-  $(".clickable-div").click(function () {
-    $("#myModal").addClass("flex").removeClass("hidden");
-  });
-});
-
-$(document).ready(function () {
-  $("#closeModal").click(function () {
-    $("#myModal").addClass("hidden").removeClass("flex");
-  });
-});
