+++ public/logout.php
