+++ public/user/qr/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>MoneyMo</title>
-    <link rel="stylesheet" href="./css/styles.css" />
+    <link rel="stylesheet" href="../../css/styles.css" />
     <link rel="icon" href="./assets/favicon.ico" type="image/x-icon" />
   </head>
 
@@ -23,7 +23,7 @@
             <div class="flex justify-center">
               <img
                 class="rounded-full w-20 h-20 border-1"
-                src="assets/default.jpg"
+                src="../../assets/default.jpg"
                 alt="image description"
               />
             </div>
@@ -38,7 +38,7 @@
             <div class="flex justify-center">
               <img
                 class="w-50 h-50"
-                src="assets/default_qr.jpg"
+                src="../../assets/default_qr.jpg"
                 alt="image description"
               />
             </div>
