+++ public/user/receipt/index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>MoneyMo</title>
-    <link rel="stylesheet" href="./css/styles.css" />
-    <link rel="icon" href="./assets/favicon.ico" type="image/x-icon" />
+    <link rel="stylesheet" href="../../css/styles.css" />
+    <link rel="icon" href="../../assets/favicon.ico" type="image/x-icon" />
   </head>
 
   <body class="bg-[#d9d9d9]">
